moving getSplits to AccumuloUtils
adding note that Socket strategy does not create splits by default
